split out some stuff to an Extras module
it's not _vital_ for querying, and opinionated, so keeping it away
